Move to Misc
Moves MoreFadePalettes and SlugcatRoomTemplates to Misc module
