+++ src/Modules/DevUIMisc/_Module.cs
@@ -13,8 +13,6 @@ public static class _Module
 	{
 		PaletteTextInput.Apply();
 		ListFixes.Apply();
-		MoreFadePalettes.Apply();
-		SlugcatRoomTemplates.Apply();
 
 		//currently used for settings saving options stuffs, but will probably later be used for much more
 		On.DevInterface.Page.ctor += Page_ctor;
@@ -29,8 +27,6 @@ public static class _Module
 	{
 		PaletteTextInput.Undo();
 		ListFixes.Undo();
-		MoreFadePalettes.Undo();
-		SlugcatRoomTemplates.Undo();
 
 		On.DevInterface.Page.ctor -= Page_ctor;
 		On.DevInterface.Page.Refresh -= Page_Refresh;
